refactor linear_coefficients
